agent_backend/pyproject.toml
[project]
name = "agent-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.127.0",
    "httpx>=0.28.1",
    "langchain>=1.2.0",
    "langchain-google-genai>=4.1.2",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.32.1",
]
